requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.23.2
httpx==0.25.0
redis==5.0.1
Pillow==10.1.0
diffusers==0.25.0
transformers==4.36.2
accelerate==0.25.0
openai==1.3.5
torch>=2.1.0
bitsandbytes==0.41.3
safetensors==0.4.1
xformers==0.0.22.post7
ninja
triton==2.1.0
flash-attn==2.3.4 